Consolidate API error types and tighten logger data
