+++ lib/__tests__/api-error.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { mapApiError } from '../../hooks/use-api-error';
 import { getApiErrorMessage } from '../api/client';
-import type { ApiError } from '../api/client';
+import type { ApiError } from '@/types/api';
 
 // ---------------------------------------------------------------------------
 // helpers
