+++ lib/api/index.ts
@@ -1,5 +1,6 @@
 export { get, post, patch, del, apiOpts } from './client';
-export type { RequestOptions, ApiError } from './client';
+export type { RequestOptions } from './client';
+export type { ApiError } from '@/types/api';
 export * as auth from './auth';
 export * as user from './user';
 export * as transfers from './transfers';
