+++ lib/logger.ts
@@ -2,7 +2,7 @@ const isDebug = process.env.NEXT_PUBLIC_DEBUG === 'true' || process.env.NODE_ENV
 
 type LogLevel = 'info' | 'warn' | 'error' | 'debug';
 
-function logMessage(level: LogLevel, message: string, data?: any) {
+function logMessage(level: LogLevel, message: string, data?: unknown) {
   if (!isDebug) return;
 
   const logEntry = {
@@ -24,8 +24,8 @@ function logMessage(level: LogLevel, message: string, data?: any) {
 }
 
 export const logger = {
-  info: (message: string, data?: any) => logMessage('info', message, data),
-  warn: (message: string, data?: any) => logMessage('warn', message, data),
-  error: (message: string, data?: any) => logMessage('error', message, data),
-  debug: (message: string, data?: any) => logMessage('debug', message, data),
+  info: (message: string, data?: unknown) => logMessage('info', message, data),
+  warn: (message: string, data?: unknown) => logMessage('warn', message, data),
+  error: (message: string, data?: unknown) => logMessage('error', message, data),
+  debug: (message: string, data?: unknown) => logMessage('debug', message, data),
 };
